64-bit DeterministicRandom() seed and getRandomSeed()
DeterministicRandom() can use a boost::random::mt19937_64
to take a 64-bit seed and return 64-bit random values,
and this simplifies things because fdb always wants 64 bits.
Performance should be slightly better because the 32-bit
mt19937 was always called twice to get 64 bits.

+++ flowbench/BenchNet2.actor.cpp
@@ -41,7 +41,7 @@ static inline TaskPriority getRandomTaskPriority(DeterministicRandom& rand) {
 ACTOR static Future<Void> benchNet2Actor(benchmark::State* benchState) {
 	state size_t actorCount = benchState->range(0);
 	state uint32_t sum;
-	state int seed = platform::getRandomSeed();
+	state uint64_t seed = platform::getRandomSeed();
 	while (benchState->KeepRunning()) {
 		sum = 0;
 		std::vector<Future<Void>> futures;
